// src/api/stetsom/model/patchApiUsersIdBody.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Stetsom API
 * API documentation for Stetsom
 * OpenAPI spec version: 1.0.0
 */
import type { UserRoleInput } from "./userRoleInput";

export type PatchApiUsersIdBody = {
  /**
   * @minLength 1
   * @maxLength 100
   */
  name?: string;
  role?: UserRoleInput;
  is_active?: boolean;
};